added my email and a github link
i added a github link in the about section and added a line below it. Also added my email to the email section because it felt empty and i wasnt sure about other peoples felling on putting their email in there. also, if you click the link for my email it will take you to you email application to compose an email